+++ Changelog.md
@@ -12,6 +12,7 @@
 * (MW) Fix SQLi detection for `$options` and `$join_conds` arguments to `SELECT` DB functions
 * (MW) Make special handling of database functions check the newer RDBMS interfaces (`IReadableDatabase` and `ISQLPlatform`)
 * (MW) Fix various crashes when using array unpack in database function calls, `getQueryInfo()` methods, or parser function hook returns
+* (MW) Restore exclusion of `Status::newGood()`, `Status::getValue()`, and `Status::setResult()` from analysis using the namespaced FQSEN.
 
 ## v6.2.1
 ### Internal changes